+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,7 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static java.util.stream.Collectors.toSet;
 
+import blackjack.domain.Denomination;
+import blackjack.domain.Suit;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Objects;
